chore(vote): reorganize files by feature
