+++ src/features/vote/proposals/list/EmptyList.astro
