+++ src/features/vote/proposals/list/ProposalList.astro
@@ -1,5 +1,5 @@
 ---
-import type { Proposal } from '../../types/proposal';
+import type { Proposal } from '../types.ts';
 import EmptyList from './EmptyList.astro';
 import LoadMoreProposalsButton from './LoadMoreProposalsButton.tsx';
 import ProposalListItem from './ProposalListItem.tsx';
